third_party/WebKit/LayoutTests/geolocation-api/permission-denied-already-error.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/geolocation-mock.js"></script>
</head>
<body>
<script>
description("Tests that when Geolocation permission has been denied prior to a call to a Geolocation method, the error callback is invoked with code PERMISSION_DENIED, when the Geolocation service encounters an error.");

var error;

geolocationMock.then(mock => {
    mock.setGeolocationPermission(false);
    mock.setGeolocationPosition(51.478, -0.166, 100);

    navigator.geolocation.getCurrentPosition(function(p) {
        testFailed('Success callback invoked unexpectedly');
        finishJSTest();
    }, function(e) {
        error = e;
        shouldBe('error.code', 'error.PERMISSION_DENIED');
        shouldBe('error.message', '"User denied Geolocation"');
        debug('');
        continueTest();
    });

    function continueTest()
    {
        // Make another request, with permission already denied.
        mock.setGeolocationPositionUnavailableError('test');

        navigator.geolocation.getCurrentPosition(function(p) {
            testFailed('Success callback invoked unexpectedly');
            finishJSTest();
        }, function(e) {
            error = e;
            shouldBe('error.code', 'error.PERMISSION_DENIED');
            shouldBe('error.message', '"User denied Geolocation"');
            finishJSTest();
        });
    }
});

window.jsTestIsAsync = true;
</script>
</body>
</html>
